Refactor ReportsView.vue adding ReportsComponent.vue and change reports to require hdid passed in for AB#15006
