+++ Apps/WebClient/src/ClientApp/src/components/report/LaboratoryReportComponent.vue
@@ -11,7 +11,6 @@ import ReportFilter from "@/models/reportFilter";
 import ReportHeader from "@/models/reportHeader";
 import { ReportFormatType, TemplateType } from "@/models/reportRequest";
 import RequestResult from "@/models/requestResult";
-import User from "@/models/user";
 import container from "@/plugins/container";
 import { SERVICE_IDENTIFIER } from "@/plugins/inversify";
 import { ILogger, IReportService } from "@/services/interfaces";
@@ -25,6 +24,9 @@ interface LabTestRow {
 
 @Component
 export default class LaboratoryReportComponent extends Vue {
+    @Prop({ required: true })
+    hdid!: string;
+
     @Prop() private filter!: ReportFilter;
 
     @Action("retrieveLaboratoryOrders", { namespace: "laboratory" })
@@ -36,19 +38,16 @@ export default class LaboratoryReportComponent extends Vue {
     @Getter("laboratoryOrdersAreLoading", { namespace: "laboratory" })
     laboratoryOrdersAreLoading!: (hdid: string) => boolean;
 
-    @Getter("user", { namespace: "user" })
-    user!: User;
-
     private logger!: ILogger;
 
     private readonly headerClass = "laboratory-test-report-table-header";
 
     private get isLaboratoryLoading(): boolean {
-        return this.laboratoryOrdersAreLoading(this.user.hdid);
+        return this.laboratoryOrdersAreLoading(this.hdid);
     }
 
     private get visibleRecords(): LaboratoryOrder[] {
-        let records = this.laboratoryOrders(this.user.hdid).filter((record) =>
+        let records = this.laboratoryOrders(this.hdid).filter((record) =>
             this.filter.allowsDate(record.timelineDateTime)
         );
         records.sort((a, b) => {
@@ -99,7 +98,7 @@ export default class LaboratoryReportComponent extends Vue {
 
     private created(): void {
         this.logger = container.get<ILogger>(SERVICE_IDENTIFIER.Logger);
-        this.retrieveLaboratoryOrders({ hdid: this.user.hdid }).catch((err) =>
+        this.retrieveLaboratoryOrders({ hdid: this.hdid }).catch((err) =>
             this.logger.error(`Error loading Laboratory data: ${err}`)
         );
     }
