+++ Apps/WebClient/src/ClientApp/src/components/report/MedicationHistoryReportComponent.vue
@@ -12,7 +12,6 @@ import ReportFilter from "@/models/reportFilter";
 import ReportHeader from "@/models/reportHeader";
 import { ReportFormatType, TemplateType } from "@/models/reportRequest";
 import RequestResult from "@/models/requestResult";
-import User from "@/models/user";
 import container from "@/plugins/container";
 import { SERVICE_IDENTIFIER } from "@/plugins/inversify";
 import { ILogger, IReportService } from "@/services/interfaces";
@@ -38,6 +37,9 @@ const options: any = {
 
 @Component(options)
 export default class MedicationHistoryReportComponent extends Vue {
+    @Prop({ required: true })
+    hdid!: string;
+
     @Prop() private filter!: ReportFilter;
 
     @Action("retrieveMedications", { namespace: "medication" })
@@ -49,16 +51,13 @@ export default class MedicationHistoryReportComponent extends Vue {
     @Getter("medications", { namespace: "medication" })
     medications!: (hdid: string) => MedicationStatementHistory[];
 
-    @Getter("user", { namespace: "user" })
-    user!: User;
-
     private logger!: ILogger;
     private notFoundText = "Not Found";
 
     private readonly headerClass = "medication-report-table-header";
 
     private get isLoading(): boolean {
-        return this.medicationsAreLoading(this.user.hdid);
+        return this.medicationsAreLoading(this.hdid);
     }
 
     private get isEmpty(): boolean {
@@ -66,7 +65,7 @@ export default class MedicationHistoryReportComponent extends Vue {
     }
 
     private get visibleRecords(): MedicationStatementHistory[] {
-        let records = this.medications(this.user.hdid).filter(
+        let records = this.medications(this.hdid).filter(
             (record) =>
                 this.filter.allowsDate(record.dispensedDate) &&
                 this.filter.allowsMedication(record.medicationSummary.brandName)
@@ -123,7 +122,7 @@ export default class MedicationHistoryReportComponent extends Vue {
 
     private created(): void {
         this.logger = container.get<ILogger>(SERVICE_IDENTIFIER.Logger);
-        this.retrieveMedications({ hdid: this.user.hdid }).catch((err) =>
+        this.retrieveMedications({ hdid: this.hdid }).catch((err) =>
             this.logger.error(`Error loading medication data: ${err}`)
         );
     }
@@ -216,7 +215,7 @@ export default class MedicationHistoryReportComponent extends Vue {
                 </template>
             </b-table>
         </section>
-        <ProtectiveWordComponent ref="protectiveWordModal" :hdid="user.hdid" />
+        <ProtectiveWordComponent ref="protectiveWordModal" :hdid="hdid" />
     </div>
 </template>
