+++ Apps/WebClient/src/ClientApp/src/components/report/NotesReportComponent.vue
@@ -10,7 +10,6 @@ import ReportFilter from "@/models/reportFilter";
 import ReportHeader from "@/models/reportHeader";
 import { ReportFormatType, TemplateType } from "@/models/reportRequest";
 import RequestResult from "@/models/requestResult";
-import User from "@/models/user";
 import UserNote from "@/models/userNote";
 import container from "@/plugins/container";
 import { SERVICE_IDENTIFIER } from "@/plugins/inversify";
@@ -24,6 +23,9 @@ interface UserNoteRow {
 
 @Component
 export default class NotesReportComponent extends Vue {
+    @Prop({ required: true })
+    hdid!: string;
+
     @Prop() private filter!: ReportFilter;
 
     @Action("retrieveNotes", { namespace: "note" })
@@ -34,9 +36,6 @@ export default class NotesReportComponent extends Vue {
 
     @Getter("notes", { namespace: "note" })
     notes!: UserNote[];
-
-    @Getter("user", { namespace: "user" })
-    private user!: User;
 
     private logger!: ILogger;
 
@@ -94,7 +93,7 @@ export default class NotesReportComponent extends Vue {
 
     private created(): void {
         this.logger = container.get<ILogger>(SERVICE_IDENTIFIER.Logger);
-        this.retrieveNotes({ hdid: this.user.hdid }).catch((err) =>
+        this.retrieveNotes({ hdid: this.hdid }).catch((err) =>
             this.logger.error(`Error loading user note data: ${err}`)
         );
     }
